Solve merge between branches
