+++ .rubocop.yml
@@ -20,7 +20,8 @@ Metrics/BlockLength:
   Max: 30
 Metrics/CyclomaticComplexity:
   Max: 20
-
+Metrics/ParameterLists:
+  Max: 10
 
 Style/Documentation:
   Enabled: false
