--- datasave/author.json
@@ -1,7 +1,0 @@
-[
-  {
-    "id": 140,
-    "first_name": "argo ",
-    "last_name": "art"
-  }
-]
