--- datasave/game.json
@@ -1,9 +1,0 @@
-[
-  {
-    "id": 630,
-    "title": "FiFi 2020",
-    "multiplayer": "yes",
-    "last_played": "05/08/2019",
-    "publish_date": "03/05/2003"
-  }
-]
