--- datasave/movie.json
@@ -1,23 +1,0 @@
-[
-  {
-    "id": 37,
-    "title": "Aron ma friend",
-    "publish_date": "14/05/2014",
-    "source": "Online",
-    "is_silent": "yes"
-  },
-  {
-    "id": 432,
-    "title": "Karla bad",
-    "publish_date": "09/08/2010",
-    "source": "Friend",
-    "is_silent": "no"
-  },
-  {
-    "id": 917,
-    "title": "hello",
-    "publish_date": "09/06/2019",
-    "source": "rudy",
-    "is_silent": "yes"
-  }
-]
