--- datasave/sources.json
@@ -1,14 +1,0 @@
-[
-	{
-		"id": "247",
-		"name": "Online"
-	},
-	{
-		"id": "796",
-		"name": "Friend"
-	},
-	{
-		"id": "415",
-		"name": "rudy"
-	}
-]
