+++ json/albums.json
@@ -1,16 +1,9 @@
 [
 	{
-		"id": "925",
-		"title": "Aron ma friend",
-		"genre": "Country",
-		"release_date": "18/02/2023",
-		"on_spotify": "true"
-	},
-	{
-		"id": "747",
-		"title": "Album2",
+		"id": "683",
+		"title": "Hello",
 		"genre": "pop",
-		"release_date": "12/08/2022",
-		"on_spotify": "true"
+		"release_date": "09/05/2008",
+		"on_spotify": "false"
 	}
 ]
