+++ json/genres.json
@@ -1,10 +1,6 @@
 [
 	{
-		"id": "112",
-		"name": "Country"
-	},
-	{
-		"id": "672",
+		"id": "554",
 		"name": "pop"
 	}
 ]
