+++ json/movie.json
@@ -1,0 +1,9 @@
+[
+  {
+    "id": 148,
+    "title": "Friend with benefit",
+    "publish_date": "05/03/2004",
+    "source": "Karla",
+    "is_silent": "yes"
+  }
+]
