+++ spec/movie_spec.rb
@@ -1,8 +1,7 @@
-require './source/source'
-require './Items/movie'
+require_relative '../properties/source'
+require_relative '../Items/movie'
 
 describe Movie do
-
   before :each do
     @source = Source.new('Online Platform')
     @movie = Movie.new('11/05/1992', 'Pearl Harbor', 'yes', @source)
@@ -23,5 +22,4 @@ describe Movie do
     @movie.move_to_archive
     expect(@movie.archived).to be true
   end
-
 end
