+++ src/app.rb
@@ -1,17 +1,12 @@
 require 'colorize'
-require './Items/movie'
-require_relative '../source/source'
-require_relative './user_input'
-require_relative './user_output'
+require_relative '../Items/movie'
+require_relative '../data-storage/user_input'
+require_relative '../data-storage/user_output'
 require_relative '../Items/game'
-require_relative '../source/author'
-require_relative './user_input'
-require_relative './user_output'
-require './Items/music_album'
-require './properties/genre'
-require './properties/source'
-require './data-storage/user_input'
-require './data-storage/user_output'
+require_relative '../properties/author'
+require_relative '../Items/music_album'
+require_relative '../properties/genre'
+require_relative '../properties/source'
 
 class App
   def initialize()
